Remove warnings, treat warning as error

+++ Runtime/Sailor.cpp
@@ -94,7 +94,7 @@ void EngineInstance::Start()
 			WCHAR Buff[50];
 			wsprintf(Buff, L"Sailor GPU FPS: %u, CPU FPS: %u", Renderer::GetInstance()->GetSmoothFps(), (uint32_t)totalFramesCount);
 			m_pInstance->m_viewportWindow.SetWindowTitle(Buff);
-			m_pInstance->m_FPS = totalFramesCount;
+			m_pInstance->m_FPS = (uint32_t)totalFramesCount;
 
 			totalFramesCount = 0;
 			totalTime = 0;
